Actualizar controladores con JWT

+++ src/main/java/com/web/proyectocanchasg1/controladores/ReservasControlador.java
@@ -1,6 +1,7 @@
 package com.web.proyectocanchasg1.controladores;
 
 import com.web.proyectocanchasg1.modelo.*;
+import com.web.proyectocanchasg1.security.keymanagment;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -27,9 +28,9 @@ public class ReservasControlador {
     public ResponseEntity<String> insertarReserva(@RequestParam String fecha,
                                                   @RequestParam int horai,
                                                   @RequestParam int horaf,
-                                                  @RequestParam String usuario,
-                                                  @RequestParam String cancha){
-
+                                                  @RequestParam String cancha,
+                                                  @RequestParam String token){
+        String usuario = keymanagment.decodeJWT(token);
         List<Canchas> c = canchasRepository.findByName(cancha);
         List<Usuario> u = usuarioRepository.findByName(usuario);
         HttpHeaders responseHeaders = new HttpHeaders();
@@ -64,24 +65,38 @@ public class ReservasControlador {
     }
 
     @CrossOrigin
-    @GetMapping(value = "/Reserva/usuario_id_usuario/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<String> getReservabyUserId(@PathVariable Long id){
-        List<Reservas> r = reservasRepository.findbyUserId(id);
-        String resResrv = r.toString();
+    @GetMapping(value = "/Reserva/usuario/{token}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<String> getReservabyUserId(@PathVariable String token){
+        String usuario = keymanagment.decodeJWT(token);
+        List<Usuario> u = usuarioRepository.findByName(usuario);
         HttpHeaders responseHeaders = new HttpHeaders();
-        return  new ResponseEntity<String>(resResrv, responseHeaders, HttpStatus.OK );
+        if(!u.isEmpty()){
+            List<Reservas> r = reservasRepository.findbyUserId(u.get(0).getIdUsuario());
+            String resResrv = r.toString();
+            return new ResponseEntity<String>(resResrv, responseHeaders, HttpStatus.OK );
+        }
+        return new ResponseEntity<String>("{\"respuesta\":\"fallo\":\"no existe usuario\"}", responseHeaders, HttpStatus.OK );
     }
 
     @CrossOrigin
     @DeleteMapping(value = "/Reserva/removeID/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<String> deleteReservasbyID(@PathVariable Long id ){
+    public ResponseEntity<String> deleteReservasbyID(@PathVariable Long id,
+                                                     @RequestParam String token){
         List<Grupo> gruposReserva = grupoRepository.findbyIdReserva(id);
         if(gruposReserva.size() == 0){
             if(reservasRepository.findById(id).isPresent()){
                 Reservas r = reservasRepository.findById(id).get();
-                reservasRepository.delete(r);
-                HttpHeaders responseHeaders = new HttpHeaders();
-                return  new ResponseEntity<String>( "{\"respuesta\":\"exito\"}", responseHeaders, HttpStatus.OK );
+                String username = keymanagment.decodeJWT(token);
+                if(r.getUsuario().getNombreU().equals(username)){
+                    reservasRepository.delete(r);
+                    HttpHeaders responseHeaders = new HttpHeaders();
+                    return  new ResponseEntity<String>( "{\"respuesta\":\"exito\"}", responseHeaders, HttpStatus.OK );
+                }
+                else{
+                    HttpHeaders responseHeaders = new HttpHeaders();
+                    return  new ResponseEntity<String>( "{\"respuesta\":\"fallo\":\"El usuario no tiene permisos sobre la reserva\"}", responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
+
+                }
             }
             else{
                 HttpHeaders responseHeaders = new HttpHeaders();
